Revert "оновлення"
