+++ src/pages/AddPetPage/Styles/chooseOption.styled.jsx
@@ -1,16 +1,19 @@
 import styled from '@emotion/styled';
 
+export const Container = styled.div``;
+
 export const CategoryForm = styled.ul`
   display: flex;
   flex-direction: column;
   align-items: flex-start;
+  width: 131px;
+  height: 176px;
   gap: 12px;
-  margin-left: 8px;
   margin-bottom: 95px;
   margin-top: 24px;
   @media screen and (min-width: 767px) {
     margin-bottom: 137px;
-    margin-top: 20px;
+    margin-top: 40px;
   }
 `;
 
@@ -30,17 +33,13 @@ export const CheckedLabel = styled.label`
 
  background-color: ${props => (props.checked ? '#54adff' : '#CCE4FB')};
   color: ${props => (props.checked ? '#FEF9F9' : '#54ADFF')};
-   :hover {
-    transition: all 0.2s ease-in-out;
-    transform: scale(1.05);
-  }
 
   > input[type='radio'] {
     // visibility: hidden;
-      border-radius: 40px;
     height: 0;
     width: 0;
   }
 
   }
 `;
+
